+++ tests/test-trashed-by.php
@@ -108,6 +108,12 @@ class Trashed_By_Test extends WP_UnitTestCase {
 		$this->assertEquals( 10, has_action( 'init', array( 'c2c_TrashedBy', 'register_meta' ) ) );
 	}
 
+	public function test_instantiating_object_fails() {
+		$this->setExpectedException( 'error' );
+
+		new c2c_TrashedBy;
+	}
+
 	public function test_meta_keys_not_created_for_post_not_trashed() {
 		$author_id = $this->create_user( false );
 		$post_id   = $this->factory->post->create( array( 'post_status' => 'draft', 'post_author' => $author_id ) );