+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * New: Add .gitignore file
 * New: Add link to plugin's page in Plugin Directory to README.md
 * Change: Register meta on `init` action instead of `plugins_loaded`
+* Change: Prevent object instantiation of the class
 * Unit tests:
     * Change: Update unit test install script and bootstrap to use latest WP unit test repo
 * Change: Note compatibility through WP 5.3+